requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Babel==3.1.0
python-dotenv==1.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.23
email_validator==2.1.0
Flask-WTF==1.1.1